requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.7
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
backports-zoneinfo==0.2.1
    # via djangorestframework
django==3.2.25
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework==3.15.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.0
    # via -r requirements.in
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pytz==2024.2
    # via django
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via
    #   asgiref
    #   pyjwt
